+++ dae/tests/integration/study_query_summary_variants/test_summary_variants_seen_in_status.py
@@ -5,7 +5,7 @@ from dae.utils.regions import Region
 from dae.variants.attributes import Status
 from dae.testing import setup_pedigree, setup_vcf, vcf_study
 
-from ...alla_import import alla_gpf
+from dae.testing.alla_import import alla_gpf
 
 
 @pytest.fixture(scope="module")
